+++ src/Controller/ServicesController.php
@@ -2,6 +2,7 @@
 
 namespace RebelCode\EddBookings\RestApi\Controller;
 
+use ArrayAccess;
 use Dhii\Data\Container\ContainerGetCapableTrait;
 use Dhii\Data\Container\ContainerHasCapableTrait;
 use Dhii\Data\Container\CreateContainerExceptionCapableTrait;
@@ -13,17 +14,24 @@ use Dhii\Exception\CreateRuntimeExceptionCapableTrait;
 use Dhii\Factory\FactoryAwareTrait;
 use Dhii\Factory\FactoryInterface;
 use Dhii\I18n\StringTranslatingTrait;
-use Dhii\Storage\Resource\SelectCapableInterface;
+use Dhii\Util\Normalization\NormalizeIntCapableTrait;
 use Dhii\Util\Normalization\NormalizeStringCapableTrait;
-use Traversable;
-use WP_Query;
+use Dhii\Util\String\StringableInterface as Stringable;
+use Dhii\Validation\Exception\ValidationFailedExceptionInterface;
+use Dhii\Validation\ValidatorInterface;
+use InvalidArgumentException;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
+use RebelCode\Entity\EntityManagerInterface;
+use stdClass;
 
 /**
  * The API controller for services.
  *
  * @since [*next-version*]
  */
-class ServicesController extends AbstractBaseCqrsController
+class ServicesController extends AbstractBaseController
 {
     /* @since [*next-version*] */
     use FactoryAwareTrait {
@@ -36,6 +44,9 @@ class ServicesController extends AbstractBaseCqrsController
 
     /* @since [*next-version*] */
     use ContainerHasCapableTrait;
+
+    /* @since [*next-version*] */
+    use NormalizeIntCapableTrait;
 
     /* @since [*next-version*] */
     use NormalizeKeyCapableTrait;
@@ -62,61 +73,55 @@ class ServicesController extends AbstractBaseCqrsController
     use StringTranslatingTrait;
 
     /**
-     * The services SELECT resource model.
+     * The default number of items to return per page.
+     *
+     * @since [*next-version*]
+     */
+    const DEFAULT_NUM_ITEMS_PER_PAGE = 20;
+
+    /**
+     * The default page number.
+     *
+     * @since [*next-version*]
+     */
+    const DEFAULT_PAGE_NUMBER = 1;
+
+    /**
+     * The services manager.
      *
      * @since [*next-version*]
      *
-     * @var SelectCapableInterface
+     * @var EntityManagerInterface
      */
-    protected $servicesSelectRm;
+    protected $servicesManager;
+
+    /**
+     * The validator for validating if a requester has access to hidden services.
+     *
+     * @since [*next-version*]
+     *
+     * @var ValidatorInterface
+     */
+    protected $serviceAuthValidator;
 
     /**
      * Constructor.
      *
      * @since [*next-version*]
      *
-     * @param SelectCapableInterface $servicesSelectRm The services SELECT resource model.
-     * @param object|null            $exprBuilder      The expression builder.
-     * @param FactoryInterface       $iteratorFactory  The iterator factory to use for the results.
+     * @param EntityManagerInterface $servicesManager      The services manager.
+     * @param FactoryInterface       $iteratorFactory      The iterator factory to use for the results.
+     * @param ValidatorInterface     $serviceAuthValidator The validator for validating if a requester has access to
+     *                                                     hidden services.
      */
     public function __construct(
-        SelectCapableInterface $servicesSelectRm,
-        $exprBuilder,
-        FactoryInterface $iteratorFactory
+        EntityManagerInterface $servicesManager,
+        FactoryInterface $iteratorFactory,
+        ValidatorInterface $serviceAuthValidator
     ) {
+        $this->servicesManager      = $servicesManager;
+        $this->serviceAuthValidator = $serviceAuthValidator;
         $this->_setIteratorFactory($iteratorFactory);
-        $this->_setExprBuilder($exprBuilder);
-        $this->servicesSelectRm = $servicesSelectRm;
-    }
-
-    /**
-     * Retrieves the services SELECT resource model.
-     *
-     * @since [*next-version*]
-     *
-     * @return SelectCapableInterface The services SELECT resource model.
-     */
-    protected function _getServicesSelectRm()
-    {
-        return $this->servicesSelectRm;
-    }
-
-    /**
-     * Sets the services SELECT resource model.
-     *
-     * @since [*next-version*]
-     *
-     * @param SelectCapableInterface $servicesSelectRm The services SELECT resource model.
-     */
-    protected function _setServicesSelectRm($servicesSelectRm)
-    {
-        if ($servicesSelectRm !== null && !($servicesSelectRm instanceof SelectCapableInterface)) {
-            throw $this->_createInvalidArgumentException(
-                $this->__('Argument is not a SELECT resource model'), null, null, $servicesSelectRm
-            );
-        }
-
-        $this->servicesSelectRm = $servicesSelectRm;
     }
 
     /**
@@ -126,32 +131,38 @@ class ServicesController extends AbstractBaseCqrsController
      */
     public function _get($params = [])
     {
-        $selectRm = $this->_getServicesSelectRm();
-
-        if ($selectRm === null) {
-            throw $this->_createRuntimeException(
-                $this->__('The services SELECT resource model is null'), null, null
-            );
+        // Get number of items per page
+        $numPerPage = $this->_containerGetDefault($params, 'numItems', static::DEFAULT_NUM_ITEMS_PER_PAGE);
+        $numPerPage = $this->_normalizeInt($numPerPage);
+        if ($numPerPage < 1) {
+            throw $this->_createControllerException($this->__('Invalid number of items per page'), 400, null, $this);
         }
 
-        $exprBuilder = $this->_getExprBuilder();
-
-        if ($exprBuilder === null) {
-            throw $this->_createRuntimeException(
-                $this->__('The SQL expression builder is null'), null, null
-            );
+        // Get page number
+        $pageNum = $this->_containerGetDefault($params, 'page', static::DEFAULT_PAGE_NUMBER);
+        $pageNum = $this->_normalizeInt($pageNum);
+        if ($pageNum < 1) {
+            throw $this->_createControllerException($this->__('Invalid page number'), 400, null, $this);
         }
 
-        $condition = $this->_buildSelectCondition($params);
+        // Calculate query offset
+        $offset = ($pageNum - 1) * $numPerPage;
 
-        // The services RM is known to require AND as top-level expression
-        if ($condition !== null) {
-            $condition = $exprBuilder->and($condition);
+        // Use the service data in params as query filters
+        $query = $this->_paramsToServiceData($params);
+
+        // Check for the `s` search term
+        if ($this->_containerHas($params, 's')) {
+            $query['s'] = $this->_containerGet($params, 's');
         }
 
-        $services  = $selectRm->select($condition);
+        try {
+            $this->serviceAuthValidator->validate($params);
+        } catch (ValidationFailedExceptionInterface $exception) {
+            $query['status'] = 'publish';
+        }
 
-        return $services;
+        return $this->servicesManager->query($query, $numPerPage, $offset);
     }
 
     /**
@@ -161,7 +172,10 @@ class ServicesController extends AbstractBaseCqrsController
      */
     protected function _post($params = [])
     {
-        return;
+        $data = $this->_paramsToServiceData($params);
+        $id   = $this->servicesManager->add($data);
+
+        return $this->_get(['id' => $id]);
     }
 
     /**
@@ -171,7 +185,19 @@ class ServicesController extends AbstractBaseCqrsController
      */
     protected function _put($params = [])
     {
-        throw $this->_createControllerException($this->__('Not implemented'), 405, null, $this);
+        $data = $this->_paramsToServiceData($params);
+
+        try {
+            $id = $this->_containerGet($data, 'id');
+        } catch (NotFoundExceptionInterface $exception) {
+            throw $this->_createControllerException(
+                $this->__('A service ID must be specified'), 400, $exception, $this
+            );
+        }
+
+        $this->servicesManager->set($id, $data);
+
+        return $this->_get(['id' => $id]);
     }
 
     /**
@@ -181,7 +207,19 @@ class ServicesController extends AbstractBaseCqrsController
      */
     protected function _patch($params = [])
     {
-        throw $this->_createControllerException($this->__('Not implemented'), 405, null, $this);
+        $data = $this->_paramsToServiceData($params);
+
+        try {
+            $id = $this->_containerGet($data, 'id');
+        } catch (NotFoundExceptionInterface $exception) {
+            throw $this->_createControllerException(
+                $this->__('A service ID must be specified'), 400, $exception, $this
+            );
+        }
+
+        $this->servicesManager->update($id, $data);
+
+        return $this->_get(['id' => $id]);
     }
 
     /**
@@ -191,62 +229,142 @@ class ServicesController extends AbstractBaseCqrsController
      */
     protected function _delete($params = [])
     {
-        throw $this->_createControllerException($this->__('Not implemented'), 405, null, $this);
+        $data = $this->_paramsToServiceData($params);
+
+        try {
+            $id = $this->_containerGet($data, 'id');
+        } catch (NotFoundExceptionInterface $exception) {
+            throw $this->_createControllerException(
+                $this->__('A service ID must be specified'), 400, $exception, $this
+            );
+        }
+
+        $this->servicesManager->delete($id);
+
+        return [];
     }
 
     /**
-     * {@inheritdoc}
+     * Extracts and creates service data from request params.
      *
      * @since [*next-version*]
+     *
+     * @param array|stdClass|ArrayAccess|ContainerInterface $params The request parameters.
+     *
+     * @return array The resulting service data.
      */
-    protected function _getSelectConditionParamMapping()
+    protected function _paramsToServiceData($params = [])
+    {
+        $mapping = $this->_getServiceDataParamMapping();
+        $data    = [];
+
+        foreach ($mapping as $_key => $_map) {
+            if (!$this->_containerHas($params, $_key)) {
+                continue;
+            }
+
+            // Get the param value
+            $_value = $this->_containerGet($params, $_key);
+            // Get the optional transformation callback
+            $_transform = isset($_map['transform']) ? $_map['transform'] : null;
+            // Transform the value
+            if ($_transform !== null) {
+                $_value = call_user_func_array($_transform, [$_value]);
+            }
+            // Get the field name
+            $_field = $_map['field'];
+            // Save in data
+            $data[$_field] = $_value;
+        }
+
+        return $data;
+    }
+
+    /**
+     * Retrieves the data param mapping.
+     *
+     * @since [*next-version*]
+     *
+     * @return array A mapping of param keys to sub-arrays that contain a `field` (the key used by the service manager)
+     *               and an optional `transform` callback for transforming a value for this field.
+     */
+    protected function _getServiceDataParamMapping()
     {
         return [
-            'id' => [
-                'compare' => 'eq',
-                'entity'  => 'service',
-                'field'   => 'id',
+            'id'              => [
+                'field'     => 'id',
+                'transform' => function ($id) {
+                    return $this->_normalizeInt($id);
+                },
+            ],
+            'name'            => [
+                'field'     => 'name',
+                'transform' => function ($name) {
+                    return $this->_normalizeString($name);
+                },
+            ],
+            'description'     => [
+                'field'     => 'description',
+                'transform' => function ($desc) {
+                    return $this->_normalizeString($desc);
+                },
+            ],
+            'status'          => [
+                'field'     => 'status',
+                'transform' => function ($status) {
+                    return $this->_normalizeString($status);
+                },
+            ],
+            'bookingsEnabled' => [
+                'field'     => 'bookings_enabled',
+                'transform' => function ($bkEn) {
+                    return $this->_normalizeInt($bkEn);
+                },
+            ],
+            'sessionLengths'  => [
+                'field' => 'session_lengths',
+            ],
+            'displayOptions'  => [
+                'field' => 'display_options',
+            ],
+            'timezone'        => [
+                'field'     => 'timezone',
+                'transform' => function ($timezone) {
+                    return $this->_normalizeString($timezone);
+                },
+            ],
+            'availability'    => [
+                'field' => 'availability',
+            ],
+            'imageId'         => [
+                'field'     => 'image_id',
+                'transform' => function ($imageId) {
+                    return $this->_normalizeInt($imageId);
+                },
             ],
         ];
     }
 
     /**
-     * {@inheritdoc}
+     * Retrieves a value from a container or data set, defaulting to a given value if not found.
      *
      * @since [*next-version*]
-     */
-    protected function _getInsertParamFieldMapping()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
      *
-     * @since [*next-version*]
-     */
-    protected function _getUpdateParamFieldMapping()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
+     * @param array|ArrayAccess|stdClass|ContainerInterface $container The container to read from.
+     * @param string|int|float|bool|Stringable              $key       The key of the value to retrieve.
+     * @param mixed                                         $default   Optional value to default to.
      *
-     * @since [*next-version*]
-     */
-    protected function _getUpdateConditionParamMapping()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
+     * @throws InvalidArgumentException    If container is invalid.
+     * @throws ContainerExceptionInterface If an error occurred while reading from the container.
      *
-     * @since [*next-version*]
+     * @return mixed The value mapped to the given key, or the $default value if the key was not found.
      */
-    protected function _getDeleteConditionParamMapping()
+    protected function _containerGetDefault($container, $key, $default = null)
     {
-        return [];
+        try {
+            return $this->_containerGet($container, $key);
+        } catch (NotFoundExceptionInterface $exception) {
+            return $default;
+        }
     }
 }